requirements.txt
# Cross-Domain Plant Classification - Herbarium to Field

torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.65.0
tensorboard>=2.13.0
PyYAML>=6.0
timm>=0.9.0
einops>=0.6.1
transformers>=4.30.0

# For EfficientNet
efficientnet-pytorch>=0.7.1

# For better logging
colorlog>=6.7.0

# For advanced augmentations
albumentations>=1.3.0

# Optional but recommended
accelerate>=0.20.0
wandb>=0.15.0
